+++ fluxo-kmp-conf/src/main/kotlin/fluxo/gradle/TaskUtils.kt
@@ -1,0 +1,3 @@
+package fluxo.gradle
+
+internal const val FLUXO_TASK_GROUP = "fluxo"