+++ fluxo-kmp-conf/src/main/kotlin/fluxo/shrink/AbstractShrinkerTask.kt
@@ -16,6 +16,7 @@ import fluxo.conf.impl.jvmToolFile
 import fluxo.conf.impl.lc
 import fluxo.external.AbstractExternalFluxoTask
 import fluxo.external.ExternalToolRunner
+import fluxo.gradle.FLUXO_TASK_GROUP
 import fluxo.gradle.cliArg
 import fluxo.gradle.ioFile
 import fluxo.gradle.ioFileOrNull